sat/requirements.txt
# Inference
einops
ignite
loguru
numpy
pandas
torch
torch_audiomentations
torchaudio

# Monitoring
sounddevice
requests
python-dotenv
tapo
